move CacheMode into separate file
